medicalStoreManagementSystem/backends/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
numpy==1.20.1
PyJWT==2.0.1
pylint==2.7.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
wrapt==1.12.1
